+++ e2e/test_range_slider_e2e.py
@@ -14,9 +14,9 @@ def _bar_page(filter_json: str = "") -> str:
 <head>
     <title>Range Slider E2E</title>
     <script src="/static/js/htmx.min.js"></script>
-    <script src="/static/js/dist/range_slider.js" type="module"></script>
-    <script src="/static/js/dist/search_select.js" type="module"></script>
-    <script src="/static/js/dist/filter_bar.js" type="module"></script>
+    <script src="/static/js/dist/elements/range-slider.js" type="module"></script>
+    <script src="/static/js/dist/elements/search-select.js" type="module"></script>
+    <script src="/static/js/dist/elements/filter-bar.js" type="module"></script>
 </head>
 <body>
     {FilterBar(filter_json=filter_json, preset_list_url="/p/l", preset_save_url="/p/s")}
