+++ tests/test_filters.py
@@ -555,8 +555,8 @@ class TestFilterBarRendering:
 
     def test_status_uses_filter_select(self):
         html = str(FilterBar())
-        assert 'data-search-select-mode="filter"' in html
-        assert 'data-name="status"' in html
+        assert 'filter-mode="true"' in html
+        assert 'name="status"' in html
 
     def test_mastered_not_checked_by_default(self):
         html = str(FilterBar(filter_json=""))
@@ -602,13 +602,13 @@ class TestFilterBarRendering:
     def test_platform_uses_search_url(self):
         """Platform is model-backed: rows are fetched, not pre-rendered."""
         html = str(FilterBar())
-        assert 'data-search-url="/api/platforms/search"' in html
+        assert 'search-url="/api/platforms/search"' in html
 
     def test_status_has_no_modifiers(self):
         """Non-nullable fields should not show (None) but MUST show (Any)."""
         html = str(FilterBar())
-        status_start = html.find('data-name="status"')
-        platform_start = html.find('data-name="platform"')
+        status_start = html.find('name="status"')
+        platform_start = html.find('name="platform"')
         status_section = html[status_start:platform_start]
         # Must have (Any) — always available
         assert "(Any)" in status_section
@@ -618,7 +618,7 @@ class TestFilterBarRendering:
     def test_platform_has_modifiers(self):
         """Nullable ForeignKey fields should show (Any)/(None)."""
         html = str(FilterBar())
-        platform_start = html.find('data-name="platform"')
+        platform_start = html.find('name="platform"')
         platform_section = html[platform_start:]
         # Should have at least one modifier option
         assert "(Any)" in platform_section or "(None)" in platform_section
